requirements.txt
libsass
jsmin
Jinja2
Psycopg2
wikipedia
Biopython
requests
bs4
gunicorn
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Script